studypal

+++ studypal/base/views.py
@@ -1,25 +1,88 @@
 from django.shortcuts import render, redirect
+from django.db.models import Q
 from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
 
 
 
 # Create your views here.
+def loginPage(request):
+
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+
+            return messages.error(request, 'User DoesNot Exist')
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+
+            return redirect('home')
+        else:
+            messages.error(request, 'username/password missing')
+
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+def logoutUser(request):
+    logout(request)
+    return redirect('home') ###
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'could not create user')
+    return render(request, 'base/login_register.html', {'form': form})
+
+
+
+
   
 def home(request):
-    rooms = Room.objects.all()
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) | #searches
+        Q(name__icontains=q) |
+        Q(description__icontains=q)
+    )
     topics = Topic.objects.all()
-    content = {'rooms': rooms, 'topics': topics}
+    room_count = rooms.count()
+    content = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', content)
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
+    room_messages = room.message_set.all()
 
-    context = {'room': room}
+    context = {'room': room, 'room_messages': room_messages}
 
     return render(request, 'base/room.html', context)
-    
+
+@login_required(login_url='login')
+
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -31,10 +94,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('you are not allowed here')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room) #instance for specific room
@@ -45,8 +111,12 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('you are not allowed here')
 
     if request.method == 'POST':
         room.delete()
